requirements.txt
streamlit
nltk
pypdf
python-docx
PyMuPDF
